refactor(eval): split evals into behavior-named files under evals/
