+++ .gitignore
@@ -44,7 +44,7 @@ coverage
 .claude
 
 # Eval debug stream dump (EVAL_DEBUG=1)
-eval/last-run.json
+evals/last-run.json
 
 # Eval secrets (API key); mirrored as a GitHub secret in CI
 .env.test.local
