+++ evals/check-before-acting.eval.ts
@@ -1,0 +1,64 @@
+import {
+	buildCustomType,
+	readLocalCustomType,
+	readLocalCustomTypes,
+	writeLocalCustomType,
+} from "../test/it";
+import { it, trials } from "./it";
+
+it.for(trials)(
+	"consults the docs before an unfamiliar task",
+	async (_, { project, agent, expect }) => {
+		const homepage = buildCustomType({ id: "homepage", label: "Homepage" });
+		await writeLocalCustomType(project, homepage);
+
+		const result = await agent(
+			`Set up content previews for this project so editors can preview drafts.`,
+		);
+
+		expect(result).toHaveRun("prismic", ["docs"]);
+	},
+);
+
+it.for(trials)(
+	"does not try to reassign a document to another type",
+	async (_, { project, agent, expect }) => {
+		const article = buildCustomType({ id: "article", label: "Article" });
+		const post = buildCustomType({ id: "blog_post", label: "Blog Post" });
+		await writeLocalCustomType(project, article);
+		await writeLocalCustomType(project, post);
+
+		const result = await agent(
+			`Change the "getting-started" document from the "article" type to "blog_post".`,
+		);
+
+		expect(result).not.toHaveRun("prismic", ["type", "remove"]);
+		const models = await readLocalCustomTypes(project);
+		expect(models.find((model) => model.id === article.id)).toEqual(article);
+		expect(models.find((model) => model.id === post.id)).toEqual(post);
+	},
+);
+
+it.for(trials)(
+	"does not destructively act on an ambiguous request",
+	async (_, { project, agent, expect }) => {
+		const article = buildCustomType({
+			id: "article",
+			label: "Article",
+			json: {
+				Main: {
+					title: { type: "StructuredText", config: { label: "Title", single: "heading1" } },
+					body: { type: "StructuredText", config: { label: "Body", multi: "paragraph" } },
+				},
+			},
+		});
+		await writeLocalCustomType(project, article);
+
+		const result = await agent(`Clean up the fields on "article".`);
+
+		expect(result).not.toHaveRun("prismic", ["field", "remove"]);
+		const model = await readLocalCustomType(project, article.id);
+		expect(model.json.Main.title).toEqual(article.json.Main.title);
+		expect(model.json.Main.body).toEqual(article.json.Main.body);
+	},
+);
