+++ evals/initialize-projects.eval.ts
@@ -1,0 +1,48 @@
+import { readFile, rm, writeFile } from "node:fs/promises";
+
+import { it, trials } from "./it";
+
+it.for(trials)(
+	"initializes Prismic in a Next.js project",
+	async (_, { project, agent, expect }) => {
+		await rm(new URL("prismic.config.json", project));
+
+		const result = await agent(`Set up Prismic in this Next.js project.`);
+
+		expect(result).toHaveRun("prismic", ["init"]);
+		const config = JSON.parse(await readFile(new URL("prismic.config.json", project), "utf8"));
+		expect(config.repositoryName).toBeTruthy();
+	},
+);
+
+it.for(trials)(
+	"adds Prismic to an existing Next.js app without clobbering it",
+	async (_, { project, agent, expect }) => {
+		await rm(new URL("prismic.config.json", project));
+		const existingPage = `export default function Home() {\n\treturn <main>KEEP-ME</main>;\n}\n`;
+		await writeFile(new URL("app/page.tsx", project), existingPage);
+
+		const result = await agent(`Add Prismic to this existing Next.js app.`);
+
+		expect(result).toHaveRun("prismic", ["init"]);
+		const config = JSON.parse(await readFile(new URL("prismic.config.json", project), "utf8"));
+		expect(config.repositoryName).toBeTruthy();
+		const page = await readFile(new URL("app/page.tsx", project), "utf8");
+		expect(page).toContain("KEEP-ME");
+	},
+);
+
+it.for(trials)(
+	"initializes with an existing repository",
+	async (_, { project, agent, expect, repo }) => {
+		await rm(new URL("prismic.config.json", project));
+
+		const result = await agent(
+			`Set up Prismic in this project using the existing "${repo}" Prismic repository.`,
+		);
+
+		expect(result).toHaveRun("prismic", ["init"]);
+		const config = JSON.parse(await readFile(new URL("prismic.config.json", project), "utf8"));
+		expect(config.repositoryName).toBe(repo);
+	},
+);
