+++ package.json
@@ -31,7 +31,7 @@
 		"types": "tsc --noEmit",
 		"unit": "cross-env MODE=test tsdown --logLevel=warn && vitest run",
 		"evals": "cross-env MODE=test RUN_EVALS=true tsdown --logLevel=warn && cross-env RUN_EVALS=true vitest run --project evals",
-		"evals:report": "node eval/report.ts",
+		"evals:report": "node evals/report.ts",
 		"unit:watch": "concurrently --raw --kill-others \"cross-env MODE=test tsdown --watch --logLevel=warn\" \"vitest watch\"",
 		"test": "npm run lint && npm run types && npm run unit"
 	},
